# Fix segfault in jody hash