@jnjackins 2021-01-25
Hi - I'm running into a similar issue with the new code. In my case I can reproduce very reliably (both in the earlier version, and this version).

```
 fatal error: checkptr: pointer arithmetic result points to invalid allocation
 
 goroutine 161 [running]:
 runtime.throw(0x11204c7, 0x40)
 	/usr/local/go/src/runtime/panic.go:1116 +0x72 fp=0xc000071340 sp=0xc000071310 pc=0x46a4d2
 runtime.checkptrArithmetic(0xc00020f510, 0xc0000713f0, 0x1, 0x1)
 	/usr/local/go/src/runtime/checkptr.go:43 +0xbe fp=0xc000071370 sp=0xc000071340 pc=0x4397be
 github.com/segmentio/fasthash/jody.AddString64(0x1c26a908921dd5e0, 0xc00020f50c, 0x4, 0x1c26a908921dd5e0)
```